feat: align the APIs into express router
